Added 6-size.c

+++ 0x00-hello_world/6-size.c
@@ -13,10 +13,10 @@ int main(void)
 	long int longIntType;
 	long long int longLongIntType;
 
-	printf("Size of a char: %lu byte(s)\n", sizeof(charType));
-	printf("Size of a int: %lu byte(s)\n", sizeof(intType));
-	printf("Size of a long int: %lu byte(s)\n", sizeof(longIntType));
-	printf("Size of a long long int: %lu byte(s)\n", sizeof(longLongIntType));
-	printf("Size of a float: %lu byte(s)\n", sizeof(floatType));
+	printf("Size of a char: %lu byte(s)\n", (unsigned long)sizeof(charType));
+	printf("Size of a int: %lu byte(s)\n", (unsigned long)sizeof(intType));
+	printf("Size of a long int: %lu byte(s)\n", (unsigned long)sizeof(longIntType));
+	printf("Size of a long long int: %lu byte(s)\n", (unsigned long)sizeof(longLongIntType));
+	printf("Size of a float: %lu byte(s)\n", (unsigned long)sizeof(floatType));
 	return (0);
 }
